+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
-
+ * [Core] Fix NPE while when failing to invoke a step definition (M.P. Korstanje)
+ 
 ## [5.1.3] (2020-01-31)
 
 ### Fixed